refactor: adjust name files folders and imports
