+++ src/app/page.tsx
@@ -4,11 +4,11 @@ import * as React from "react";
 import { useTheme } from "next-themes";
 import Link from "next/link";
 import { GoogleAnalytics } from "@next/third-parties/google";
-import Header from "../components/Header/Header";
-import { PlayIcon } from "../components/Icons/playIcon";
-import { MainButton } from "../components/PrimaryButton";
-import { LeadBoardIcon } from "../components/Icons/leadboardIcon";
-import { HowToPlayIcon } from "../components/Icons/howPlay";
+import Header from "@/components/Header/Header";
+import { PlayIcon } from "@/components/Icons/playIcon";
+import { MainButton } from "@/components/PrimaryButton";
+import { LeadBoardIcon } from "@/components/Icons/leadboardIcon";
+import { HowToPlayIcon } from "@/components/Icons/howPlay";
 import { Icon } from "@/components/Icons/icon";
 
 export default function HomePage() {
@@ -35,13 +35,25 @@ export default function HomePage() {
         </p>
         <div className="flex gap-4 flex-col w-[280px]">
           <Link href="/game" className="w-full">
-            <MainButton title="Play Now" icon={<PlayIcon />} className="bg-button-main text-white py-3" /> 
+            <MainButton
+              title="Play Now"
+              icon={<PlayIcon />}
+              className="bg-button-main text-white py-3"
+            />
           </Link>
           <div className="w-full">
-            <MainButton title="Leaderboard" icon={<LeadBoardIcon />} className="bg-button-normal border-2 border-solid text-black py-3" />
+            <MainButton
+              title="Leaderboard"
+              icon={<LeadBoardIcon />}
+              className="bg-button-normal border-2 border-solid text-black py-3"
+            />
           </div>
           <div className="w-full">
-            <MainButton title="How to Play?" icon={<HowToPlayIcon />} className="bg-button-normal border-2 border-solid text-black py-3"/>
+            <MainButton
+              title="How to Play?"
+              icon={<HowToPlayIcon />}
+              className="bg-button-normal border-2 border-solid text-black py-3"
+            />
           </div>
         </div>
       </div>
