+++ src/components/Footer/FooterButton.tsx
@@ -1,4 +1,4 @@
-import { MainButton } from "../PrimaryButton";
+import { MainButton } from "@/components/PrimaryButton";
 
 export default function FooterButton() {
   return (
@@ -9,7 +9,10 @@ export default function FooterButton() {
       <div className="w-52">
         <hr className="h-px my-8 bg-gray-200 border-0 dark:bg-gray-700" />
         <div>
-          <MainButton title="Continue" className="text-button-main border border-solid-100 gap-x-2 px-4 py-2 bg-button-main text-white py-3"/>
+          <MainButton
+            title="Continue"
+            className="text-button-main border border-solid-100 gap-x-2 px-4 py-2 bg-button-main text-white py-3"
+          />
         </div>
       </div>
     </div>
