+++ src/components/Icons/howPlay.tsx
