+++ src/components/Icons/logoSolo.tsx
