+++ src/components/Icons/playIcon.tsx
