+++ src/components/TextInput.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Input } from "../ui/input";
+import { Input } from "./ui/input";
 import Textarea from "@mui/joy/Textarea";
